+++ src/subapps/projects/components/AddComponentButton.less
@@ -1,7 +1,7 @@
 @import '../../../shared/lib.less';
 
 .add-component-button {
-  background-color: rgb(247, 245, 245);
+  background-color: @primary-color;
   border-radius: 10px;
   box-shadow: @box-shadow-base;
   width: 70px;
@@ -9,7 +9,6 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  // temporary until we update canvas
   position: fixed;
   z-index: 21;
   margin: 15px;
