+++ src/subapps/projects/components/StatusIcon.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 
+import { Status } from '../types';
+
 const progressIcon = require('../../../shared/images/progressIcon.svg');
 const blockedIcon = require('../../../shared/images/blockedStatus.svg');
 const todoIcon = require('../../../shared/images/todoStatus.svg');
@@ -7,13 +9,6 @@ const doneIcon = require('../../../shared/images/done.svg');
 const progressIconMini = require('../../../shared/images/progressIconMini.svg');
 const blockedIconMini = require('../../../shared/images/blockedIconMini.svg');
 const doneIconMini = require('../../../shared/images/doneIconMini.svg');
-
-export enum Status {
-  toDo = 'not started',
-  inProgress = 'in progress',
-  blocked = 'blocked',
-  done = 'done',
-}
 
 const StatusIcon: React.FC<{
   status: Status;
