+++ src/subapps/projects/containers/WorkflowStepsContainer.tsx
@@ -6,12 +6,12 @@ import { NexusClient } from '@bbp/nexus-sdk';
 import SingleStepContainer from './SingleStepContainer';
 import StepsBoard from '../components/WorkflowSteps/StepsBoard';
 import { displayError, successNotification } from '../components/Notifications';
-import { StepResource, WorkflowStepMetadata } from '../views/WorkflowStepView';
 import ProjectPanel from '../components/ProjectPanel';
 import { fetchTopLevelSteps } from '../utils';
 import AddComponentButton from '../components/AddComponentButton';
 import WorkflowStepWithActivityForm from '../components/WorkflowSteps/WorkflowStepWithActivityForm';
 import fusionConfig from '../config';
+import { StepResource, WorkflowStepMetadata } from '../types';
 
 const WorkflowStepContainer: React.FC<{
   orgLabel: string;
